backend/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.4.0
python-decouple==3.8
